+++ app/smartbrain_3000/web/_app/immutable/chunks/D5hYGZ4w.js
@@ -1,0 +1,1 @@
+import"./C-1FJjmk.js";